# docs/api-documentation.yml
openapi: 3.0.3
info:
  title: Embedded Systems Design Platform API
  description: |
    A comprehensive REST API for the AI-Embedded Systems Design Platform.
    This platform enables users to design, simulate, and deploy embedded systems projects
    with AI-powered assistance and real-time collaboration features.
  version: 1.0.0
  contact:
    name: API Support
    email: support@embedded-platform.com
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.embedded-platform.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: Create a new user account with email, username, and password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
                - name
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                  description: Unique username (letters, numbers, underscores only)
                  example: "johndoe"
                email:
                  type: string
                  format: email
                  description: Valid email address
                  example: "john.doe@example.com"
                password:
                  type: string
                  minLength: 8
                  description: Password with at least 8 characters, 1 uppercase, 1 lowercase, 1 number
                  example: "SecurePass123"
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  description: Full display name
                  example: "John Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT authentication token
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: User login
      description: Authenticate user with email/username and password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  description: Email address or username
                  example: "john.doe@example.com"
                password:
                  type: string
                  description: User password
                  example: "SecurePass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT authentication token
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update user profile
      description: Update the authenticated user's profile information
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 50
                  example: "John Doe Updated"
                bio:
                  type: string
                  maxLength: 500
                  example: "Embedded systems enthusiast"
                website:
                  type: string
                  format: uri
                  example: "https://johndoe.dev"
                location:
                  type: string
                  maxLength: 100
                  example: "San Francisco, CA"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete user account
      description: Deactivate the authenticated user's account
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Account deactivated successfully"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        username:
          type: string
          example: "johndoe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        name:
          type: string
          example: "John Doe"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        avatar:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            bio:
              type: string
              maxLength: 500
              example: "Embedded systems enthusiast"
            website:
              type: string
              format: uri
              example: "https://johndoe.dev"
            location:
              type: string
              example: "San Francisco, CA"
            isActive:
              type: boolean
              example: true
            emailVerified:
              type: boolean
              example: true
            lastLogin:
              type: string
              format: date-time
              example: "2023-12-01T10:30:00.000Z"
            stats:
              type: object
              properties:
                projectsCreated:
                  type: integer
                  example: 15
                projectsCompleted:
                  type: integer
                  example: 12
                totalLogins:
                  type: integer
                  example: 45
                lastActivity:
                  type: string
                  format: date-time
                  example: "2023-12-01T10:30:00.000Z"
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email is required"
              value:
                type: string
                example: "invalid-email"